refactor: move packages around
