+++ adventofcode.common/src/main/java/com/lewisbirks/adventofcode/common/tuple/Pair.java
@@ -1,0 +1,4 @@
+package com.lewisbirks.adventofcode.common.tuple;
+
+public record Pair<L,R>(L left, R right) {
+}
