--- adventofcode.common/src/main/java/com/lewisbirks/adventofcode/tuple/Pair.java
@@ -1,4 +1,0 @@
-package com.lewisbirks.adventofcode.tuple;
-
-public record Pair<L,R>(L left, R right) {
-}
