+++ adventofcode.year2021/src/main/java/com/lewisbirks/adventofcode/day/Day17.java
@@ -2,7 +2,7 @@ package com.lewisbirks.adventofcode.day;
 
 import com.lewisbirks.adventofcode.common.cache.CachedSupplier;
 import com.lewisbirks.adventofcode.common.domain.Day;
-import com.lewisbirks.adventofcode.model.point.Point;
+import com.lewisbirks.adventofcode.common.coor.Point;
 
 import java.util.function.Supplier;
