+++ adventofcode.year2021/src/main/java/com/lewisbirks/adventofcode/day/Day20.java
@@ -1,7 +1,7 @@
 package com.lewisbirks.adventofcode.day;
 
 import com.lewisbirks.adventofcode.common.domain.Day;
-import com.lewisbirks.adventofcode.model.point.Point;
+import com.lewisbirks.adventofcode.common.coor.Point;
 
 import java.util.ArrayList;
 import java.util.BitSet;
