+++ adventofcode.year2021/src/main/java/com/lewisbirks/adventofcode/day/Day21.java
@@ -1,6 +1,6 @@
 package com.lewisbirks.adventofcode.day;
 
-import com.lewisbirks.adventofcode.collection.FrequencyMap;
+import com.lewisbirks.adventofcode.common.collection.FrequencyMap;
 import com.lewisbirks.adventofcode.common.cache.CachedSupplier;
 import com.lewisbirks.adventofcode.common.domain.Day;
