+++ adventofcode.year2021/src/main/java/com/lewisbirks/adventofcode/model/Scanner.java
@@ -1,6 +1,6 @@
 package com.lewisbirks.adventofcode.model;
 
-import com.lewisbirks.adventofcode.collection.FrequencyMap;
+import com.lewisbirks.adventofcode.common.collection.FrequencyMap;
 import com.lewisbirks.adventofcode.model.point.Point3D;
 import com.lewisbirks.adventofcode.utils.ReferenceInt;
