+++ adventofcode.year2022/src/main/java/com/lewisbirks/adventofcode/day/Day4.java
@@ -2,7 +2,7 @@ package com.lewisbirks.adventofcode.day;
 
 import com.lewisbirks.adventofcode.common.cache.CachedSupplier;
 import com.lewisbirks.adventofcode.common.domain.Day;
-import com.lewisbirks.adventofcode.tuple.Pair;
+import com.lewisbirks.adventofcode.common.tuple.Pair;
 
 import java.util.List;
 import java.util.function.Supplier;
